@@ -4,13 +4,14 @@ version: 0.0.2
 homepage: https://github.com/ajinasokan/autopilot
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_test:
     sdk: flutter
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
 
